requirements.txt
pandas
tk
python-docx
